plugins/enthusiast-model-google/pyproject.toml
[tool.poetry]
name = "enthusiast-model-google"
version = "0.1.0"
description = ""
authors = ["Damian Sowiński <damian.sowinski@upsidelab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
enthusiast-common = "~=0.6"
google-genai = "^1.18.0"
langchain-google-genai = "^2.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
